+++ NEWS.md
@@ -1,3 +1,8 @@
+# cookies 0.2.2
+
+* Updated cookie handling to allow HttpOnly cookies mixed with JavaScript cookies, and improved error messages (#54).
+* Updated cookie handling to find cookies even if the call comes from inside a module (#56).
+
 # cookies 0.2.1
 
 * The underlying 'js-cookie' JavaScript library now loads from source if it is available (#42).